+++ test/src/widgets/reactive_checkbox_test.dart
@@ -77,7 +77,7 @@ void main() {
         expect(checkbox.value, false);
 
         // When: set to True the form control
-        form.formControl('isChecked').value = true;
+        form.control('isChecked').value = true;
         await tester.pump();
 
         // Then: the checkbox is checked
@@ -103,7 +103,7 @@ void main() {
         expect(checkbox.value, true);
 
         // When: set to False the form control
-        form.formControl('isChecked').value = false;
+        form.control('isChecked').value = false;
         await tester.pump();
 
         // Then: the checkbox is checked