+++ test/src/widgets/reactive_radio_test.dart
@@ -70,7 +70,7 @@ void main() {
         await tester.pumpWidget(ReactiveRadioTestingWidget(form: form));
 
         // When: changes control to true
-        form.formControl('radio').value = true;
+        form.control('radio').value = true;
         await tester.pump();
 
         // Expect radio group value is true
@@ -99,7 +99,7 @@ void main() {
         await tester.pumpWidget(ReactiveRadioTestingWidget(form: form));
 
         // When: changes control to false
-        form.formControl('radio').value = false;
+        form.control('radio').value = false;
         await tester.pump();
 
         // Expect radio group value is true